[o11y] Fix unused WorkerTracer warnings in R2
With #5685, we managed to reduce the volume of "destructed WorkerTracer"
warnings by >80% by eliminating it in the case of duplicate alarm events.
Another case where this happens are R2 API calls where we create a new
WorkerInterface for the R2 call before completing error checking, as seen in the
r2-test itself. We can easily avoid this by moving getHttpClient() calls behind
error checks.
While this PR is not aiming for completion I also cleaned up a call in
web-socket.c++ that may be susceptible to the same issue.