Run alarm without syncing alarm manager if currentTime past local time
Historically, if a SQLite DOs local alarm time in SQLite didn't match
the AlarmManager's alarm time, we would reschedule the alarm and try to
run the alarm again later.

However, if the time we are attempting to sync to is already before the
current real system time, there is no point in rescheduling to run
later, since we might as well run now.

This commit runs the alarm immediately if the current system time is
already past the time we are trying to sync the alarm manager to, since
it's better to run an alarm a little late rather than very late.